Projeto Uolcompass/sprints-4-5-pb/src/app/requirements.txt
fastapi
uvicorn[standard]
boto3
joblib
python-dotenv
numpy
pydantic
xgboost